+++ R/cvar-package.R
@@ -4,6 +4,8 @@
 #' @importFrom gbutils cdf2quantile
 #' @importFrom Rdpack reprompt
 #'
+#' @importFrom fGarch dstd pstd qstd rstd dged pged qged rged
+#'
 #' @docType package
 #' @name cvar-package
 #' @aliases cvar